requirements.txt
transformers==4.27.3
pypng
Pillow